02/requirements.txt
apache-airflow~=2.1.0
requests